geoToMem accuracy improvements
Changed geoToMem to use distance instead of dot product to determine
increment. Scales this difference inversely to the total distance from
current to goal. Current implementation has a * 10 to account for a
factor difference. This is likely an approximation of a parameter yet to
be identified, and may not hold for files of all resolutions. Indeed, we
still see out-of-ops, with error as high as 1.0 on different
invocations. This should be investigated. However, current geoToMem can
reach perfect accuracy (i.e., same pixel as SeaDAS) on 09GA files.