Per-site water masks and geoToMem2 OOB handling
Made it actually load all the independent site water masks. geoToMem2
now tracks depth to avoid a stack overflow and to determine when the
point is likely to be out of bounds. This method is cheaper and less
trouble than detecting improper convergence, but if there are issues
with accuracy the allowed depth may need to be dialed in further.